% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACESO.R
\docType{package}
\name{ACESO}
\alias{ACESO}
\alias{ACESO-package}
\title{ACESO}
\description{
ACESO: A Cancer Evolution Simulation Optimizer
}
\details{
ACESO is a multiscale computational framework to explore
the evolutionary dynamics of heterogeneous tumor cell populations taking
pharmacokinetic and drug interaction effects into account to rationally identify
optimum single-agent and combination treatment administration strategies for
oncogene-driven cancers.
}
\keyword{Cancer}
\keyword{Modeling&Simulation}
\keyword{Stochastic-models}
\keyword{branching-process}

