% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Alpelisib_sensitive}
\alias{Alpelisib_sensitive}
\title{Alpelisib_sensitive dataset}
\usage{
data(Alpelisib_sensitive)
}
\description{
Dataset containing the effect of alpelisib on the birth and death rates of alpelisib-sensitive BT-20 cell line (example data adapted from the effect of alpelisib on the BT-20 cell line data obtained from the HMS LINCS database)..
}
\details{
The data consists of a dataframe named Alpelisib_sensitive with the following columns:
 \itemize{
  \item Cell.line: name of the cell line used in the experiment. BT-20 in this example.
  \item Time: time point (hours).
  \item CONC: concentration of the drug 1. Numeric column.
  \item Type: drug resistant cell type. Number 1 in this case.
  \item Net_growth: net growth rates of the cells under different drug concentrations.
  \item Birth_rate: birth rate of the sensitive cell line influenced by different drug concentrations.
  \item Death_rate: Constant death rate.
}
}
\examples{
\dontrun{
 data(Alpelisib_sensitive)
 head(Alpelisib_sensitive)
}
}
\keyword{datasets}
