% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_death_rate.R
\name{calculate_death_rate}
\alias{calculate_death_rate}
\title{calculate_death_rate}
\usage{
calculate_death_rate(cell_death_data, net_growth_data = NULL,
  d_start = 0.005, Apoptotic.fraction = T,
  column.name = "Apoptotic.fraction", ...)
}
\arguments{
\item{cell_death_data}{name of the dataframe where the apoptosis assay data is saved.}

\item{net_growth_data}{name of the dataframe where the growth kinetics data is saved.}

\item{d_start}{initial parameter value for the death rate to be used in the least squared estimation process.}

\item{Apoptotic.fraction}{if TRUE, the fraction of death cells is used for the calculation of the death rate. If FALSE, the total number of death cells is used.}

\item{column.name}{string value to specify the column name where the total number or the fraction of death cells numbers are saved.}

\item{...}{Additional arguments for nls function}
}
\value{
A dataframe where the death rate for each cell type is included as a column to the information used as input.
}
\description{
Function to calculate the death rate from the apoptosis assay data.
}

