% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Dactolisib_Trametinib_combination}
\alias{Dactolisib_Trametinib_combination}
\title{Dactolisib_Trametinib_combination dataset}
\usage{
data(Dactolisib_Trametinib_combination)
}
\description{
Dataset containing the effect of the combination of Dactolisib and Trametinib on the number of BT-20 cell line cell counts.
}
\details{
The data consists of a dataframe named Dactolisib_Trametinib_combination with the following columns:
 \itemize{
  \item Cell.line: name of the cell line used in the experiment. BT-20 in this example.
  \item Small.Molecule.Name: Name of one of the drug used in the experiment. Dactolisib in this case.
  \item CONC: concentration of the drug specified in Small.Molecule.Name column.
  \item CONC.units: units of drug concentrations in CONC. Numeric column.
  \item Small.Molecule.2.Name: Name of the second drug used in the experiment. Trametinib in this case.
  \item CONC2: concentration of the drug specified in Small.Molecule.2.Name column. Numeric column.
  \item CONC2.units: units of drug concentrations in CONC2.
  \item Biological.Replicate: only 1 in this example. Numeric column.
  \item Replicate: Technical Replicate. Numeric column.
  \item Cell_Count_0: Total cell counts before treatment.
  \item Viable.cells: Total cell count after treatment
  \item Control: Total control cell count.
  \item Normalized_GR: Normalized Growth Rate Inhibition value.
  \item Time: Time where the viable cells are counted after treatment.
  \item Type: drug sensitive cell type. Number 0 in this case.
  }
}
\examples{
\dontrun{
 data(Dactolisib_Trametinib_combination)
 print(Dactolisib_Trametinib_combination)
}
}
\keyword{datasets}

