% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions_u.R
\name{define.Type0.cells}
\alias{define.Type0.cells}
\title{define.Type0.cells}
\usage{
define.Type0.cells(N0, birth_rate, death_rate, scale = 1,
  pk.function = "pk")
}
\arguments{
\item{N0}{cell population at time 0.}

\item{birth_rate}{birth rate function. It can be a numeric value, a user created function or the result of a model fitting function.}

\item{death_rate}{death rate function. It can be a numeric value, a user created function or the result of a model fitting function.}

\item{scale}{scaling parameter. Numeric value.}

\item{pk.function}{the name of the function(s) which computes the pharmacokinetic profile of the drug(s) that affects type 0 cells.}
}
\value{
a S4 object with all the above information gathered together.
}
\description{
Define Type 0 cells: sensitive cell lines
}
\examples{
define.Type0.cells(N0=1,birth_rate = 0.001,death_rate  = 0.005)
}
