% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions_u.R
\name{define.Typei.cells}
\alias{define.Typei.cells}
\title{define.Typei.cells}
\usage{
define.Typei.cells(Ni = 0, birth_rate, death_rate, mutation_rate,
  scale = 1, pk.function = "pk")
}
\arguments{
\item{Ni}{cell population at time 0.}

\item{birth_rate}{birth rate function. It can be a numeric value, a user created function or the result of a model fitting function.}

\item{death_rate}{death rate function. It can be a numeric value, a user created function  or the result of a model fitting function.}

\item{mutation_rate}{numeric value or function for the mutation rate.}

\item{scale}{scaling parameter. Numeric value.}

\item{pk.function}{the name of the function which computes the pharmacokinetic profile of the drug that affects type i cells.}
}
\value{
a S4 object with all the above information gathered together.
}
\description{
Define Type i cells: resistant cell line
}
\examples{
define.Typei.cells(Ni=0,birth_rate = 0.001,death_rate  = 0.005,mutation_rate=0.0001)
}
