% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_functions_u.R
\name{define.Typeicr.cells}
\alias{define.Typeicr.cells}
\title{define.Typeicr.cells}
\usage{
define.Typeicr.cells(Ni_0 = 0, birth_rate, death_rate, mutation_rate,
  mother_cell = 1, scale = 1)
}
\arguments{
\item{Ni_0}{cell population at time 0.}

\item{birth_rate}{numeric value for the birth rate of cross-resistant cells.}

\item{death_rate}{numeric value for the death rate of cross-resistant cells.}

\item{mutation_rate}{numeric value for mutation rate.}

\item{mother_cell}{number corresponding to the mother cell type from which type icr cell has emerged.}

\item{scale}{scaling parameter. Numeric value.}
}
\value{
a S4 object with all the above information gathered together.
}
\description{
Define Type icr cells: cross-resistant cell line
}
\examples{
define.Typeicr.cells(Ni_0=0,birth_rate = 0.001,death_rate  = 0.005,mutation_rate=0.0001)

}
