% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response_surface.R
\name{DifferenceSurface.plot}
\alias{DifferenceSurface.plot}
\title{DifferenceSurface.plot}
\usage{
DifferenceSurface.plot(rmap, zcenter = NULL, xl = expression(CONC),
  yl = expression(CONC2), zl = "Effect", low = "#e9a3c9",
  mid = "#f7f7f7", high = "#4dac26", title = "")
}
\arguments{
\item{rmap}{a ggplot2 plot object depicting a response map, produced by the function responseMap.}

\item{zcenter}{the value that forms the centerpoint (white) in the bidirectional color scale. If NULL (the default), the default setting for scale_fill_gradient2 is used.}

\item{xl}{The character string or expression used to label the x-axis}

\item{yl}{The character string or expression used to label the y-axis}

\item{zl}{The character string or expression used to label the colorscale representing the response variable. "Effect" by default.}

\item{low}{character corresponding to the color selected for the lowest values in the response map.}

\item{mid}{color for the mid point.}

\item{high}{character corresponding to the color selected for the highest values in the response map.}

\item{title}{title for the plot}
}
\value{
A ggplot2 plot object, with the appropriate scales and labels added.
}
\description{
DifferenceSurface.plot
}
\seealso{
braidReports::responseMap
}

