% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PK.R
\name{easy.mrgsim}
\alias{easy.mrgsim}
\title{easy.mrgsim}
\usage{
easy.mrgsim(model, dosing_schedule = NULL, parameters = NULL,
  variability = NULL, precision = 0.1, tend, scale = 1,
  missing_times = NULL, ...)
}
\arguments{
\item{model}{Structural PK model definition (mrgsolve model specification)}

\item{dosing_schedule}{Dosing schedule to simulate the PK model}

\item{parameters}{Estimates of the pharmacokinetic parameters of the model.}

\item{variability}{variability terms for the specified PK parameters.}

\item{precision}{number: increment of the sequence to simulate from 0 to tend.}

\item{tend}{Final time for the simulation of the PK model.}

\item{scale}{scaling parameter}

\item{missing_times}{numeric vector to indicate the times where the patient missed some of the doses specified in the dosing_schedule argument.}

\item{...}{additional arguments for the mrgsim function from mrgsolve package}
}
\value{
This function computes predictions of PK models following mrgsolve model specifications.
}
\description{
Extended version of mrgsim function from mrgsolve package to compute predictions of PK models following mrgsolve model specifications.
}
\examples{
\dontrun{
# mrgsolve model specification for a one compartment model and intravenuos administration:
 code_iv <- '$PARAM @annotated
 TVCL   :  8 : Clearance (volume/time)
 TVV    : 80 : Central volume (volume)
 
 $CMT  @annotated
CENT : Central compartment


$MAIN
double CL = exp(log(TVCL) + ETA1);
double V = exp(log(TVV)  + ETA2);

$OMEGA @labels ETA1 ETA2
0 0

$GLOBAL
#define CP (CENT)

$PKMODEL ncmt = 1, depot = FALSE

$CAPTURE @annotated
CP : Plasma concentration (mass/volume)'

#Read the code:
cmt1_iv<- mcode("mymodel", code_iv)\%>\% Req(CP) 

#Specify dosing schedule: 150mg once a day for a month (30 days). The dose is given at time 0.
d1 <- ev(amt = 150, ii = 1, addl = 30, time=0)

#Define or change the parameters for the model. Use a list:
newpar <- list(TVCL=7.14*24,TVV=155)
# * TVCL: Typical value for the clearance (volume/time) 
# * TVV: Typical value for the volumen of distribution
# To see the names of the parameters in the model:
param(cmt1_iv)

#Simulate the model with easy.mrgsim function 
easy.mrgsim(model=cmt1_iv,dosing_schedule=d1,tend=30,delta=0.1,parameters = newpar) \%>\% plot

#Now simulate another dosign schedule: 1600mg once a week for 45 days. The dose is given at day 2.
d2 <- ev(amt = 1600, ii = 7, addl = 30, time=2)

easy.mrgsim(model=cmt1_iv,dosing_schedule=d2,tend=45,delta=0.1,parameters = newpar) \%>\% plot

#Save the results as a data.frame
Simulation2<-as.data.frame(
easy.mrgsim(model=cmt1_iv,dosing_schedule=d2,tend=45,delta=0.1,parameters = newpar))
}
}

