% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_resistance.R
\name{En_crossresistant_cells_i}
\alias{En_crossresistant_cells_i}
\title{En_crossresistant_cells_i}
\usage{
En_crossresistant_cells_i(t, type_icr, type_0, type_i, i,
  approximation = T)
}
\arguments{
\item{t}{Final time}

\item{type_icr}{Type-icr S4 object}

\item{type_0}{Type-0 S4 object}

\item{type_i}{list with Type-i S4 object}

\item{i}{cell type number from which the cross-resistance is originating}

\item{approximation}{logical argument indicating if an approximation of the numerical integration method must be used or not. Default to TRUE for faster computation.}
}
\value{
En_crossresistant_cells_i returns the number of cross-resistant cells that arise from resistant clone i at time t after treatment initiation
}
\description{
Expected number of cross-resistant cells arising from resistant clone i at time t after treatment initiation
}
