% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/En_resistant_cells_i.R
\name{En_resistant_cells_i}
\alias{En_resistant_cells_i}
\title{En_resistant_cells_i}
\usage{
En_resistant_cells_i(t, type_0, type_i, i, type_icr = NULL,
  approximation = T)
}
\arguments{
\item{t}{Final time}

\item{type_0}{Type-0 S4 object}

\item{type_i}{list with Type-i S4 object}

\item{i}{number of cell type}

\item{type_icr}{Type-icr S4 object for the i cell type}

\item{approximation}{logical argument indicating if an approximation of the numerical integration method must be used or not. Default to TRUE for faster computation.}
}
\value{
En_resistant_cells returns the number of resistant cells in clone i at time t after treatment initiation
}
\description{
Expected number of resistant cells in clone i at time t after treatment initiation
}
\examples{
\dontrun{
#Birth rate of type 0 cells as a function of time:
b0=function(time){0.05*sin(0.1*time)+0.1}
#Create Type-0 S4 object structure with the parameters of type 0 sensitive cells
Type0 <-define.Type0.cells(N0=100,birth_rate = b0,death_rate= 0.14)
#Birth rate of type 1 cells as a function of time:
b1=function(time){0.05*sin(0.1*time)+0.12}
#Create Type-i S4 object structure with the parameters of type i resistant cells
Type1 <- define.Typei.cells(Ni=0,birth_rate = b1,death_rate  = 0.09,mutation_rate=10^-3)
#Group the different type i resistant cells
type_i<-list(Type1)
#  #Call En_resistant_cells_i function
ni_t<-En_resistant_cells_i(t=100,type_0=Type0,type_i=type_i,i=1,approximation=F)
}
}
