% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN_type0_cells.R
\name{EN_type0_cells}
\alias{EN_type0_cells}
\title{EN_type0_cells}
\usage{
EN_type0_cells(T0 = 0, t, type_0, ui = 0, int.function = c("integrate",
  "pracma"))
}
\arguments{
\item{T0}{Initial time. Default 0.}

\item{t}{Final time}

\item{type_0}{Type-0 S4 object}

\item{ui}{vector containing the mutation probabilities of the resistant cell clones. Default is 0 (single-type birth-death process, no resistant clones).}

\item{int.function}{integration function. Possible options are "integrate" or "pracma". The default option is integrate function in R. If "pracma" is selected, the numerical integration methods from pracma package are used (more robust but slower option).}
}
\value{
EN_type0_cells returns the number of sensitive cells at time t after treatment initiation
}
\description{
Expected number of sensitive type 0 cells at time t after treatment initiation
}
\examples{
\dontrun{
#Birth rate of type 0 cells as a function of time:
b0=function(time){0.05*sin(0.1*time)+0.1}
#Create Type-0 S4 object structure with the parameters of type 0 sensitive cells
Type0 <-define.Type0.cells(N0=1,birth_rate = b0,death_rate= 0.08)
#Call EN_type0_cells function to calculate the number of cells at time 250.
EN_type0_cells(t=250,type_0=Type0)
EN_type0_cells(T0=0,t=50,type_0=Type0)*EN_type0_cells(T0=50,t=250,type_0=Type0)
}

}

