% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PK.R
\name{Estimate.PK}
\alias{Estimate.PK}
\title{Estimate.PK}
\usage{
Estimate.PK(PK_data, x.name = "time", y.name = "CONC", model,
  initial_estimates, log.yaxis = FALSE, weighted = FALSE)
}
\arguments{
\item{PK_data}{Data to fit (drug concentrations over time)}

\item{x.name}{Column name of the independent variable (time) in the x axis. "time" by default.}

\item{y.name}{Column name of the dependent variable (drug concentrations) in the y axis. "CONC" by default.}

\item{model}{Structural PK model definition (mrgsolve model specification).}

\item{initial_estimates}{list with the initial estimates for the PK parameters.}

\item{log.yaxis}{logical to indicate if the values for the dependent variable should be log transformed. FALSE by default.}

\item{weighted}{logical. If TRUE a weighted NLS method is applied.}
}
\value{
This function returns the final estimates for the PK paramters and the plot of the observation values and the prediction of the model
}
\description{
Fit Pharmacokinetic (PK) data and estimate PK parameters. The use of other software like NONMEM or MONOLIX is encoraged for estimation of parameters from different individuals.
}
\examples{
\dontrun{
data(PK_example)
head(PK.example.data)
#Read code from model repository: 1 compartment model with extravascular administration
cmt1_oral<- mread("PK_models/1cmt_oral")\%>\% Req(CP) 
#See parameter names in the model:
param(cmt1_oral)
#Estimate parameters:

}
}
