% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GD}
\alias{GD}
\title{Dactolisib_Trametinib_rates dataset}
\usage{
data(Dactolisib_Trametinib_rates)
}
\description{
Dataset containing the effect of the combination of Dactolisib and Trametinib on the birth rate of drug sensitive BT-20 cell line.
}
\details{
The data consists of a dataframe named GD with the following columns:
 \itemize{
  \item Cell.line: name of the cell line used in the experiment. BT-20 in this example.
  \item CONC: concentration of the drug 1. Numeric column.
  \item CONC2: concentration of the drug 2. Numeric column.
  \item Type: drug sensitive cell type. Number 0 in this case.
  \item Net_growth: net growth rates of the cells under different drug concentrations.
  \item Birth_rate: birth rate of the sensitive cell line influenced by different drug concentrations.
  \item Death_rate: Constant death rate.
}
}
\examples{
\dontrun{
 data(Dactolisib_Trametinib_rates)
 head(GD)
}
}
\keyword{datasets}
