% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSA.R
\name{HSA}
\alias{HSA}
\title{Fill the dataset with the predicted effect under Highest Single Agent (HSA) model}
\usage{
HSA(data, resp = "Birth_rate", conc1 = "CONC", conc2 = "CONC2",
  Emax.effect = c("min", "max"))
}
\arguments{
\item{data}{input dataset}

\item{resp}{Name of the column in the dataset with the effect/response to be fitted.}

\item{conc1}{Name of the column in the dataset with the concentrations for drug 1. 'CONC' by default.}

\item{conc2}{Name of the column in the dataset with the concentrations for drug 1. 'CONC2' by default.}

\item{Emax.effect}{argument to indicate whether the maximum effect of the drug in the dataset is associated with the minimum or the maximum values of the resp column.}
}
\value{
The input dataset with a new column called 'HSA_response' to store the expected effect of combination under HSA model
}
\description{
Fill the dataset with the predicted effect under Highest Single Agent (HSA) model
}
\examples{
NULL
}

