% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fitting.R
\name{linear.fit}
\alias{linear.fit}
\title{linear.fit}
\usage{
linear.fit(data, resp = "Birth_rate", conc = "CONC", ...)
}
\arguments{
\item{data}{Concentration-effect dataframe.}

\item{resp}{Name of the column with the response values. Default is Birth_rate.}

\item{conc}{Name of the column with the drug concentration values.}

\item{...}{Additional arguments for the lm function.}
}
\value{
This function returns the results of the model fitting function
}
\description{
Function to fit linear models using lm function.
}
