% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loewe.R
\name{Loewe}
\alias{Loewe}
\title{Fill the dataset with the predicted effect under Loewe model}
\usage{
Loewe(data, resp = "Birth_rate", conc1 = "CONC", conc2 = "CONC2",
  same.Emax = F, Emax.effect = c("min", "max"))
}
\arguments{
\item{data}{input data}

\item{resp}{Name of the column in the dataset with the effect/response to be fitted.}

\item{conc1}{Name of the column in the dataset with the concentrations for drug 1. 'CONC' by default.}

\item{conc2}{Name of the column in the dataset with the concentrations for drug 1. 'CONC2' by default.}

\item{same.Emax}{logical value to indicate whether the hill model for drug 1 and drug 2 share the same Emax value (classical Loewe model). FALSE by default.}

\item{Emax.effect}{argument to indicate whether the maximum effect of the drug in the dataset is associated with the minimum or the maximum values of the resp column.}
}
\value{
The input dataset with a new column called 'loewe_additivity' to store the expected effect of combination under Loewe additive model
}
\description{
Fill the dataset with the predicted effect under Loewe model
}
\examples{
\dontrun{
data(Dactolisib_Trametinib_rates)
head(GD)
GD=Loewe(data=GD,resp = 'Birth_rate')
head(GD)
rmap <-  braidReports::responseMap(loewe_additivity~CONC+CONC2,GD,logscale=T)#' 
plot.ResponseSurface(rmap,xl="[Dactolisib] (µM)",yl="[Trametinib] (µM)",
zl="Birth rate of \\n sensitive cells (1/h)",title="Loewe")
}
}

