% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fitting.R
\name{model.select}
\alias{model.select}
\title{model.select}
\usage{
model.select(formula, data, fctList = NULL, type = "continuous",
  nested = FALSE, sorted = c("IC", "Res var", "Lack of fit", "no"),
  icfct = AIC, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. Either of the form 'response ~ dose' or as a data frame with response values in first column and dose values in second column.}

\item{data}{Concentration-response dataframe.}

\item{fctList}{a list of dose-response functions to be compared.}

\item{type}{a character string specifying the data type (parameter estimation will depend on the data type as different log likelihood function will be used).}

\item{nested}{logical. TRUE results in F tests between adjacent models (in 'fctList'). Only sensible for nested models.}

\item{sorted}{character string determining according to which criterion the model fits are ranked.}

\item{icfct}{function supplying the information criterion to be used. "AIC" and "BIC" are the two options. "AIC" by default.}

\item{...}{Additional arguments for the model fitting function. See \code{\link[drc]{drm}} for more info.}
}
\value{
This function compares different models using the following criteria: the log likelihood value, Akaike's or Bayes information criterion (AIC) or the estimated residual standard error.
}
\description{
Comparison of different models using the following criteria: the log likelihood value, Akaike's or Bayes information criterion (AIC) or the estimated residual standard error.
}
