% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_library.R
\name{model_library}
\alias{model_library}
\title{Internal model library from ACESO}
\usage{
model_library(list = FALSE)
}
\arguments{
\item{list}{logical:list available models}
}
\description{
Internal model library from ACESO
}
\details{
Call \code{model_library(list=TRUE)} to list available models.  Once the model 
is loaded (see examples below), call \code{(mod)@code} to see
model code and equations.
}
\examples{
\dontrun{
model_library(list=TRUE)
mod <- mread("1cmt_iv", model_library())
mod
mod <- mread("1cmt_ev", model_library()) 
mod@code

} 
}

