% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fitting.R
\name{Multiple.best.singlefit}
\alias{Multiple.best.singlefit}
\title{Multiple.best.singlefit}
\usage{
Multiple.best.singlefit(data, resp = "Birth_rate", conc = "CONC",
  IC = "AIC", type = "continuous", ...)
}
\arguments{
\item{data}{Concentration-response dataframe.}

\item{resp}{Response to be fitted (Y axis). Default is Birth_rate.}

\item{conc}{Name of the column with the drug concentration values.}

\item{IC}{string for supplying the information criterion to be used. "AIC" and "BIC" are the two options. "AIC" by default.}

\item{type}{a character string specifying the data type (parameter estimation will depend on the data type as different log likelihood function will be used).}

\item{...}{Additional arguments for the selection of the best model fitting function. See ?model.select for more info.}
}
\value{
This function returns the best model fitted to each cell line data.
}
\description{
Function where linear and several non-linear models are used to fit the curve of single drug effects on multiple cell lines at the same time.
}
\examples{
\dontrun{
#Effect of Alvocidib on the cell counts of BT-20 and MCF7 cell lines.

filename=system.file("extdata","2cell_lines.txt",package="ACESO")

growth_data=read.cellcount.data(filename,sep="\\t")

#Calculate net growth rate assuming an exponential growth of the cells:
growth_data<-net_growth_rate(growth_data)

#Fit the data with a 4 parameter log-logistic function:
Multiple.best.singlefit(growth_data,resp="Net_growth",conc="CONC")
}
}

