% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_growth_rate.R
\name{net_growth_rate}
\alias{net_growth_rate}
\title{net_growth_rate}
\usage{
net_growth_rate(Inputdata, death_rate = NULL, birth_rate = NULL,
  time0_data = T)
}
\arguments{
\item{Inputdata}{name of the dataframe where the cell viability data is saved.}

\item{death_rate}{numeric value for the death rate. Default to NULL.}

\item{birth_rate}{numeric value for the birth rate. Default to NULL.}

\item{time0_data}{boolean argument to specify if the cell viability data for time 0 values should be used or not. Default to TRUE.}
}
\value{
The Inputdata dataframe with the net growth rate values for each cell type added.
}
\description{
Function to calculate the net growth rate from the cell viability assay data.
}

