% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_extinction.R
\name{Pext}
\alias{Pext}
\title{Pext}
\usage{
Pext(t, Text, type_i, i, approximation = T, int.function = c("integrate",
  "pracma"))
}
\arguments{
\item{t}{time of production of a resistant cell clone}

\item{Text}{time of extinction}

\item{type_i}{Type-i S4 object}

\item{i}{resistant clone identifier}

\item{approximation}{logical argument indicating if an approximation of the numerical integration method must be used or not. Default to TRUE for faster computation.}

\item{int.function}{integration function. Possible options are "integrate" or "pracma". The default option is integrate function in R. If "pracma" is selected, the numerical integration methods from pracma package are used (more robust but slower option).}
}
\value{
Pext function returns the probability of extinction (between 0 and 1) at time T of a single resistant cell produced at time t
}
\description{
Probability of extinction at time Text of a single resistant cell produced at time t
}

