% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probability_extinction.R
\name{Pext_preexisting_resist}
\alias{Pext_preexisting_resist}
\title{Pext_preexisting_resist}
\usage{
Pext_preexisting_resist(t, type_i, i)
}
\arguments{
\item{t}{time of production of a resistant cell clone}

\item{type_i}{Type-i S4 object}

\item{i}{resistant clone identifier}
}
\value{
Pext_preexisting_resist function returns the probability of extinction (between 0 and 1) at time t of a single resistant cell produced at time t
}
\description{
Probability of extinction at time t of a resistant cell type when there is pre-existing resistance at treatment initiation
}

