% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PK.example}
\alias{PK.example}
\title{PK_example dataset}
\usage{
data(PK.example)
}
\description{
Plasma drug concentration over time for 1 individual.
}
\details{
The data consists of a dataframe named PK.example where the plasma drug concentrations (mg/L, CONC column) over time (hours, time column)  for one individual (ID column) are stored.
 The individual receives a dose of 150mg (amt column) at time 0.
}
\examples{
\dontrun{
 data(PK.example)
 print(PK.example)
}
}
\keyword{datasets}

