% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PK.R
\name{pk.function}
\alias{pk.function}
\title{pk.function}
\usage{
pk.function(model, dosing_schedule = NULL, parameters = NULL,
  variability = NULL, tend, delta = 5e-05, pk.function = NULL,
  missing_times = NULL, output.time = NULL, output.conc = NULL,
  scale = 1, scale.x = 1, ...)
}
\arguments{
\item{model}{Structural PK model definition (mrgsolve model specification)}

\item{dosing_schedule}{Dosing schedule to simulate the PK model}

\item{parameters}{Estimates of the pharmacokinetic parameters of the model.}

\item{variability}{variability terms for the specified PK parameters.}

\item{tend}{Final time for the simulation of the PK model.}

\item{delta}{a number indicating the increment of a time sequence to simulate from 0 to tend. A small value is needed for a proper interpolation. Default: 0.00005.}

\item{pk.function}{user defined pk function to interpolate instead of interpolating the simulation results of a specified PK model}

\item{missing_times}{numeric vector to indicate the times where the patient missed some of the doses specified in the dosing_schedule argument.}

\item{output.time}{vector of times to interpolate}

\item{output.conc}{vector of drug concentrations to interpolate}

\item{scale}{scaling parameter for y axis}

\item{scale.x}{scaling parameter for x axis}

\item{...}{additional arguments for the mrgsim function from mrgsolve package}
}
\value{
This function returns an interpolation of PK data or a simulated PK model
}
\description{
Makes an interpolation of Pharmacokinetic (PK) data or a simulated PK model over time
}
\examples{
\dontrun{
# mrgsolve model specification for a one compartment model and intravenuos administration:
 code_iv <- '$PARAM @annotated
 TVCL   :  8 : Clearance (volume/time)
 TVV    : 80 : Central volume (volume)
 
 $CMT  @annotated
CENT : Central compartment


$MAIN
double CL = exp(log(TVCL) + ETA1);
double V = exp(log(TVV)  + ETA2);

$OMEGA @labels ETA1 ETA2
0 0

$GLOBAL
#define CP (CENT)

$PKMODEL ncmt = 1, depot = FALSE

$CAPTURE @annotated
CP : Plasma concentration (mass/volume)'

library(mrgsolve)

#Read the code:
cmt1_iv<- mcode("mymodel", code_iv)\%>\% Req(CP) 

#Specify dosing schedule: 150mg once a day for a month (30 days)
e1 <- ev(amt = 150, ii = 1, addl = 30, time=0)

#Define or change the parameters for the model. Use a list:
newpar <- list(TVCL=7.14*24,TVV=155)
# * TVCL: Typical value for the clearance (volume/time) 
# * TVV: Typical value for the volumen of distribution

#Pk function to be used during the simulation of the evolutionary process:
pk=pk.function(model=cmt1_iv,dosing_schedule=e1,tend=30,parameters = newpar)

}
}
