% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.cellcount.data.R
\name{read.cellcount.data}
\alias{read.cellcount.data}
\title{read.cellcount.data}
\usage{
read.cellcount.data(filename, sep = ",", combination = F)
}
\arguments{
\item{filename}{name of the csv file. string.}

\item{sep}{the field separator character. Comma by default.}

\item{combination}{boolean argument to specify if the file has drug combination data, that is, more that one drug concentration column. Default to FALSE.}
}
\value{
A dataframe with the relevant information about the cell viability assay.
}
\description{
Read the csv file where the cell viability assay data is gathered.
}
