% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_death_rate.R
\name{read.celldeath.file}
\alias{read.celldeath.file}
\title{read.celldeath.file}
\usage{
read.celldeath.file(filename, sep = ",", combination = F,
  column.name = "Apoptotic.fraction")
}
\arguments{
\item{filename}{name of the csv file. string.'Time', 'CONC' and the column name specified
in column.name argument are mandatory columns in the file.}

\item{sep}{the field separator character. Comma by default.}

\item{combination}{boolean argument to specify if the file has drug combination data, that is, more that one drug concentration column. Default to FALSE.}

\item{column.name}{string value to specify the column name where the total number or the fraction of death cells numbers are saved.}
}
\value{
A dataframe with the relevant information about the apoptosis assay.
}
\description{
Read the csv file where the apoptosis assay data is saved.
}
