% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PK.R
\name{read.PKdata}
\alias{read.PKdata}
\title{read.PKdata}
\usage{
read.PKdata(file)
}
\arguments{
\item{file}{name of the file where to drug concentrations over time are stored.}
}
\value{
A data.frame with the pharmacokinetic data
}
\description{
Read the pharmacokinetic data
}

