% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response_surface.R
\name{ResponseSurface.plot}
\alias{ResponseSurface.plot}
\title{ResponseSurface.plot}
\usage{
ResponseSurface.plot(rmap, palette = NULL, cscale = NULL,
  xl = expression(CONC), yl = expression(CONC2), zl = "Effect",
  title = "")
}
\arguments{
\item{rmap}{a ggplot2 plot object depicting a response map, produced by the function responseMap.}

\item{palette}{a character vector or two or more colors defining the colorscale to be applied to the fill variable in the plot.}

\item{cscale}{a value around which the colorscale applied to the fill variable will be centered. If NULL (the default), the colorscale will range from the minimum fill variable to the maximum.}

\item{xl}{The character string or expression used to label the x-axis}

\item{yl}{The character string or expression used to label the y-axis}

\item{zl}{The character string or expression used to label the colorscale representing the response variable. "Effect" by default.}

\item{title}{title for the plot}
}
\value{
A ggplot2 plot object, with the appropriate scales and labels added.
}
\description{
ResponseSurface.plot
}
\seealso{
braidReports::responseMap
}

