% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loewe.R
\name{SynergyMatrix.plot}
\alias{SynergyMatrix.plot}
\title{Colored concentration pairwise synergy matrix}
\usage{
SynergyMatrix.plot(data, resp = "diffLoewe", conc1 = "CONC",
  conc2 = "CONC2")
}
\arguments{
\item{data}{input data}

\item{resp}{Name of the column in the dataset with the effect/response to be fitted.}

\item{conc1}{Name of the column in the dataset with the concentrations for drug 1. 'CONC' by default.}

\item{conc2}{Name of the column in the dataset with the concentrations for drug 1. 'CONC2' by default.}
}
\value{
Colored concentration pairwise matrix reflecting the difference between the measurement and the surface obtained under a no-interaction model. Values less than zero (blue) represent antagonism and values greater that zero (yellow) represent synergism.
}
\description{
Colored concentration pairwise synergy matrix
}
\examples{
\dontrun{
data(Dactolisib_Trametinib_rates)
head(GD)
GD=Loewe(data=GD,resp = 'Birth_rate')
GD$diffLoewe=(GD$loewe_additivity-GD$Birth_rate)
SynergyMatrix.plot(GD, resp="diffLoewe")
}
}
