% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Trametinib_sesitive.data}
\alias{Trametinib_sesitive.data}
\title{Trametinib_sensitive.data dataset}
\usage{
data(Trametinib_sesitive.data)
}
\description{
Dataset containing the effect of Trametinib on the birth and death rates of trametinib-sensitive BT-20 cell line 
(example data adapted from the effect of trametinib on the BT-20 cell line data obtained from the HMS LINCS database).
}
\details{
The data consists of a dataframe named Trametinib_sensitive.data with the following columns:
 \itemize{
  \item Cell.line: name of the cell line used in the experiment. BT-20 in this example.
  \item Time: time point (hours).
  \item CONC: concentration of the drug 1. Numeric column.
  \item Type: drug resistant cell type. Number 2 in this case.
  \item Net_growth: net growth rates of the cells under different drug concentrations.
  \item Birth_rate: birth rate of the sensitive cell line influenced by different drug concentrations.
  \item Death_rate: Constant death rate.
}
}
\examples{
\dontrun{
 data(Trametinib_sesitive.data)
 head(Trametinib_sesitive.data)
}
}
\keyword{datasets}
