% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loewe.R
\name{try.loewe}
\alias{try.loewe}
\title{Compute the predicted effect under Loewe model}
\usage{
try.loewe(d1, d2, model, same.Emax = F)
}
\arguments{
\item{d1}{dose of drug 1}

\item{d2}{dose of drug 2}

\item{model}{hill model for drug 1 and drug 2}

\item{same.Emax}{logical value to indicate whether the hill model for drug 1 and drug 2 share the same Emax value (classical Loewe model)}
}
\value{
expected effect of combination under Loewe additive model
}
\description{
Compute the predicted effect under Loewe model
}
\examples{
NULL
}

