% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_fitting.R
\name{Multiple.best.singlefit}
\alias{Multiple.best.singlefit}
\title{Multiple.best.singlefit}
\usage{
Multiple.best.singlefit(data, resp = "Net_growth", conc = "CONC",
  IC = "AIC", type = "continuous", ...)
}
\arguments{
\item{data}{Concentration-response dataframe.}

\item{resp}{Response to be fitted (Y axis). Default is Net_growth.}

\item{conc}{Name of the column with the drug concentration values.}

\item{IC}{string for supplying the information criterion to be used. "AIC" and "BIC" are the two options. "AIC" by default.}

\item{type}{a character string specifying the data type (parameter estimation will depend on the data type as different log likelihood function will be used).}

\item{...}{Additional arguments for the selection of the best model fitting function. See ?model.select for more info.}
}
\value{
This function returns the best model fitted to each cell line data.
}
\description{
Function where linear and several non-linear models are used to fit the curve of single drug effects on multiple cell lines at the same time.
}

